+++ src/lib/libc/_libc.c
@@ -1,11 +1,22 @@
 #include "_libc.h"
 
 #include <stdio.h>
-#include "k/syscalls.h"
+#include <k/syscalls.h>
 
 void __libc_init()
 {
-    // TODO : when pipes, update
+    // --- stdio --- //
+    // Std streams
+    stdin = malloc(sizeof(FILE));
+    stdout = malloc(sizeof(FILE));
+    stderr = malloc(sizeof(FILE));
+
+    stdin->_fileno = 0;
+    stdout->_fileno = 1;
+    stderr->_fileno = 2;
+
+
+    // TODO : rm
     // Bind stdin to libc stdin
-    strcon(stdin, __libc_stdinCallback);
+    strcon(0, __libc_stdinCallback);
 }