+++ src/lib/libc/_libc.h
@@ -6,10 +6,15 @@
 #include <stddef.h>
 
 // --- Init --- //
+// TODO : change to start
 // Used to init the library
 // Call this on app entry
 void __libc_init();
 
+// TODO : Desturctors (stdio files)
+
+
+// TODO : rm
 // --- stdin --- //
 // When data comes to stdin this function is called
 void __libc_stdinCallback(void *stdinStream, uint8_t *buf, size_t n);