+++ src/lib/libc/stdio.h
@@ -1,13 +1,27 @@
 #pragma once
 
+#include <k/types.h>
+
+#define EOF (-1)
+
+typedef struct FILE_t
+{
+    // File descriptor
+    fd_t _fileno;
+} FILE;
+
 // Std streams
-#define stdin 0
-#define stdout 1
-#define stderr 2
+extern FILE *stdin;
+extern FILE *stdout;
+extern FILE *stderr;
+
+int fputc(int c, FILE *stream);
 
 char *gets(char *s);
 
 int getchar();
+
+#define putc fputc
 
 int putchar(int c);