+++ src/lib/libk/k/io.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#define STDIN_FD 0
+#define STDOUT_FD 1
+#define STDERR_FD 2