+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 - Added color app to change the color of the tty
 - Added Virtual File System (vfs)
 - Added Fat32 driver
+- Added the k library, syscalls + types...
 - Moved libc and kernel/k to lib
 - Remade syscalls
 - Added data structures like vector and queue to k
 - Added pipes in kernel
+- Added app entry / termination (Ctrl + C works)
 
 ## 0.1